+++ obras/.obras_utils.sh
@@ -10,9 +10,9 @@
 
 
 # variables
-export OBRAS_UTILS_VERSION=1.5.23
+export OBRAS_UTILS_VERSION=1.5.24
 export OBRAS_UTILS_VERSION_DATE=2021.04.06
-export OBRAS_UTILS_UPDATE_MESSAGE="Correct 'site services start' when there is no tmp."
+export OBRAS_UTILS_UPDATE_MESSAGE="Correct 'redis' and 'pip3' installations."
 
 export OS=`uname`
 if [ $OS == 'Darwin' ]; then
@@ -170,27 +170,30 @@ obras_utils() {
         fi
       fi
 
-      if ! test -f /usr/local/bin/pipx && ! test -f $HOME/.local/bin/pipx ; then
+      if ! test -f /usr/local/bin/pipx && ! test -f /usr/bin/pip3; then
         echo -e "\033[1;92m==> \033[0m\033[1;39mInstalling \"pip\" \033[0m"
         echo ""
         if [ "$OS" == 'Darwin' ]; then
           brew install pipx
           pipx ensurepath
         else 
-          python3 -m pip install --user pipx
-          python3 -m pipx ensurepath 
-          source $HOME/.bashrc
+          sudo apt-get install python3-pip
         fi
       fi
 
-      if ! test -f $HOME/.local/bin/iredis; then
+      if ! test -f $HOME/.local/bin/iredis && ! test -f /usr/local/bin/iredis; then
         echo -e "\033[1;92m==> \033[0m\033[1;39mInstalling \"iredis\" \033[0m"
         echo ""
         if [ "$OS" == 'Darwin' ]; then
           brew install iredis
         else 
-          sudo apt-get install python3-venv
-          pipx install iredis
+          test -f ./iredis.tar.gz && rm -f ./iredis.tar.gz
+          test -f ./iredis && rm -f ./iredis
+          test -d lib && rm -rf lib
+          wget "https://github.com/laixintao/iredis/releases/latest/download/iredis.tar.gz" && test -f ./iredis.tar.gz && tar -xzf ./iredis.tar.gz && sudo mv ./iredis /usr/local/bin && sudo mv lib /usr/local/bin
+          test -f ./iredis.tar.gz && rm -f ./iredis.tar.gz
+          test -f ./iredis && rm -f ./iredis
+          test -d lib && rm -rf lib
         fi
       fi
